btree: land forceBusySnapshotForTest behind btreetesthooks tag
SQLite gates fault-injection helpers behind SQLITE_TEST; mirror that
with a build-tag const so the hook check compiles away in default
builds (verified: beginWriteWithSnapshot 165 vs 173 insns). Un-skips
the two BeginWrite ErrBusySnapshot retry-contract tests, previously
commented-out husks. NOTES drift-131 documents the mechanism.
