+++ internal/btree/test_hooks_on.go
@@ -1,0 +1,5 @@
+//go:build btreetesthooks
+
+package btree
+
+const walTestHooks = true
