+++ internal/btree/wal.go
@@ -1427,12 +1427,10 @@ type wal struct {
 	// If nil, lock failures return ErrBusy immediately (issue 1.7).
 	busyHandler BusyHandler
 
-	// Test hook — intentionally commented out to avoid a per-BeginWrite
-	// atomic.Load on the production hot path. Uncomment together with the
-	// matching check in beginWriteWithSnapshot to re-enable the two tests
-	// in db_test.go (TestBeginWrite_SurfacesBusySnapshotAfterBoundedRetries,
-	// TestBeginWrite_BusySnapshotRoutesThroughBusyHandler).
-	// forceBusySnapshotForTest atomic.Bool
+	// forceBusySnapshotForTest makes beginWriteWithSnapshot fail with
+	// ErrBusySnapshot on every attempt. Checked only under
+	// -tags btreetesthooks (walTestHooks); compiles away otherwise.
+	forceBusySnapshotForTest atomic.Bool
 
 	// writerHdr is the writer's private copy of the SHM header, updated after
 	// each successful commit (writeFrames) and after re-sync in beginWrite().
@@ -3069,12 +3067,9 @@ func (w *wal) beginWrite() (stateChanged bool, err error) {
 // beginWriteWithSnapshot is the generalized form: caller supplies the
 // read snapshot for the BUSY_SNAPSHOT check. See beginWrite for semantics.
 func (w *wal) beginWriteWithSnapshot(readSnap WalIndexHdr) (stateChanged bool, err error) {
-	// Test hook — kept commented to avoid an atomic.Load on every BeginWrite.
-	// Uncomment with the matching field in the wal struct to run the two
-	// ErrBusySnapshot tests in db_test.go.
-	// if w.forceBusySnapshotForTest.Load() {
-	// 	return false, ErrBusySnapshot
-	// }
+	if walTestHooks && w.forceBusySnapshotForTest.Load() {
+		return false, ErrBusySnapshot
+	}
 	if err := walBusyLock(w.index, w.busyHandler, lockWrite, lockExclusive); err != nil {
 		return false, err
 	}
